feat: [prd-backend-package-restructure-002] - [Extract worker subprocess plumbing into a dedicated package]
