+++ docs/internal/processes/development-guide-relevant-files.md
@@ -143,6 +143,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/factory/runtime/dispatch_result_hook_test.go` | Dispatch-hook regression surface | Planner-present synchronous execution, planner-absent async worker-pool submission, missing-runner failures, and replay-tick validation should stay covered here when the behavior being changed is local to `workerPoolDispatchResultHook`, instead of forcing every dispatch-hook regression through the broader factory integration suite. |
 | `pkg/factory/subsystems/subsystem_dispatcher.go` | Dispatcher snapshot-enrichment owner | Dispatcher execution must reattach `d.state` onto raw runtime snapshots before handing them to scheduler/enablement flows, because runtime-only snapshots intentionally omit topology while runtime guards still need normalized transition-to-worker mappings. |
 | `pkg/workers/event_exit_code.go` | Shared worker event exit-code emission policy | Script and inference boundary-event payloads should route their final exit-code inclusion decision through this package-private seam so zero-vs-nonzero emission rules stay centralized even when diagnostic extraction differs by caller. |
+| `pkg/workers/process/` and `pkg/workers/command.go` | Worker subprocess boundary split | Subprocess request shaping, environment merging, timeout and exit handling, and platform-specific command execution now belong under `pkg/workers/process`, while `pkg/workers/command.go` stays as the stable compatibility seam for broadly consumed worker command types. Future worker package moves should preserve this pattern: move implementation into the narrow subpackage first, then keep or trim the root boundary deliberately instead of letting execution mechanics drift back into `pkg/workers`. |
 | `pkg/workers/template_fields.go` | Workstation parameterized field-resolution owner | Working-directory, environment-variable, and worktree template resolution for workstation execution should stay centralized here so runtime cleanup work updates one explicit helper contract instead of reintroducing call-shape shims or duplicating template evaluation in `workstation_executor.go` or tests. |
 | `pkg/workers/workstation_executor_test.go` | Workstation execution-context regression owner | When worker runtime-field cleanup changes how workstation templates are applied, keep the observable regression proof here by asserting the dispatched runtime context or `resolveWorkstationExecutionContext` output values rather than adding signature-only or helper-ownership tests elsewhere. |
 | `pkg/workers/provider_behavior.go` and `pkg/workers/content_capability.go` | Provider CLI behavior and content-capability owner | Provider-specific argument construction and command-request shaping live in `provider_behavior.go`; worker inference paths should call this seam instead of reintroducing forwarding wrappers in `inference_provider.go`. Codex canonical image content should become ordered `-i <file>` arguments here after validating file readability against the subprocess working directory, while non-Codex text-only execution paths should reuse the shared content-capability helper to fail image-bearing inputs before subprocess invocation. |
