+++ pkg/workers/process/command_process_unix.go
@@ -1,6 +1,6 @@
 //go:build !windows
 
-package workers
+package process
 
 import (
 	"os/exec"
